test proofed

+++ ft_memset.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_memset.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tmanolis <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: tmanolis <tmanolis@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/05/28 11:18:19 by tmanolis          #+#    #+#             */
-/*   Updated: 2021/05/28 14:15:56 by tmanolis         ###   ########.fr       */
+/*   Created: 2021/06/22 17:18:53 by tmanolis          #+#    #+#             */
+/*   Updated: 2021/06/22 17:48:36 by tmanolis         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	*ft_memset(void *b, int c, size_t len)
 {
 	unsigned char	*str;
-	int	i;
+	size_t	i;
 
 	str = (unsigned char	*)b;
 	i = 0;
